+++ src/components/POLine/ConfirmReceivingWorkflowChange/index.js
@@ -1,0 +1,1 @@
+export { ConfirmReceivingWorkflowChangeModal } from './ConfirmReceivingWorkflowChangeModal';