+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Format the "Publishers" column to avoid displaying nullish values. Fixes UIOR-1436.
 * Display an error toast only when PO Line locations violate fund restrictions. Fixes UIOR-1420.
 * Display the fiscal year in which an order was opened. Fixes UIOR-1403.
+* Automatically set to `Independent order and receipt quantity` if a user selects `Receipt not required` for opened order. UIOR-1243.
 
 ## [8.0.4](https://github.com/folio-org/ui-orders/tree/v8.0.4) (2025-06-12)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v8.0.3...v8.0.4)